\name{fastPSSIM-package}
\alias{fastPSSIM-package}
\alias{fastPSSIM}
\docType{package}
\title{\packageTitle{fastPSSIM}}
\description{\packageDescription{fastPSSIM}}
\section{Package Content}{\packageIndices{fastPSSIM}}
\author{\packageAuthor{fastPSSIM}}
\section{Maintainer}{\packageMaintainer{fastPSSIM}}
\keyword{package}
