% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{makepseudo}
\alias{makepseudo}
\title{Nearest neighbor augmentation based on ranks}
\usage{
makepseudo(N, n, k, a, alltrt)
}
\arguments{
\item{N}{Total number of covariate values.}

\item{n}{Vector of sample sizes from all treatments.}

\item{k}{Number of nearest neighbors.}

\item{a}{Number of treatment levels in the data.}

\item{alltrt}{A matrix of dimension 3x\code{N}, with first two rows being Y 
and X, and a third row with the rank of the X values within the same 
treatment level.}
}
\value{
A list containing the following:
\itemize{
  \item \strong{psudo}: A 3D array of dimensions (\code{k}, \code{a}, 
  \code{N}) that stores the augmented observations based on the k-nearest 
  neighbor rule in Wang, Tolos and Wang (2010).
  \item \strong{index}: A 3D array of dimensions (\code{k}, \code{a}, 
  \code{N}) that stores the indices of the observations used for 
  augmentation.
}
}
\description{
The function \code{makepseudo()} performs the nearest neighbor augmentation 
based on the rank of covariate values according to the scheme described on 
page 410-411 of Wang, Tolos and Wang (2010). 

This is a rewrite of the original \code{makepseudo()} function from the 
PSSIM library using Rcpp to decrease runtime. If you load both this library 
and PSSIM, load this one second so the faster functions will overwrite the 
slower ones.
}
\examples{
 a = 2; n = c(7, 9); N = sum(n);  X = runif(N);
 trt = c(rep(1, n[1]), rep(2, n[2])); e = rnorm(N, 0, 0.1)
 Y = ifelse(trt == 1, 4*(X-0.5)^2+e, 2*X+e)
 ranksuse = unlist(tapply(X, trt, rank))
 alltrt = rbind(Y, X, ranksuse)
 aug = makepseudo(N, n, k=3, a, alltrt)
 
}
\references{
Haiyan Wang, Siti Tolos, and Suojin Wang (2010). A Distribution Free
 Nonparametric Test to Detect Dependence Between a Response Variable and
 Covariate in Presence of Heteroscedastic Treatment Effects.
 The Canadian Journal of Statistics. 38(3), 408433. Doi:10.1002/cjs.10068
}
\seealso{
\code{\link[=NPtest_indept]{NPtest_indept}} uses this function to 
run the independence test.
}
