% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mapindex}
\alias{mapindex}
\title{Index in one vector mapped to treatment and observation index}
\usage{
mapindex(r, n)
}
\arguments{
\item{r}{An integer between 1 and \code{num(n)}.}

\item{n}{A vector of the sample sizes.}
}
\value{
A 2-d index, where the first gives which treatment \code{r} belongs 
to and the second gives which observation \code{r} is within that treatment.
}
\description{
Function \code{mapindex()} maps the 1-d index \eqn{r = 1, \ldots, N} to 2-d 
index \eqn{i = 1, \ldots, a}, \eqn{j = 1, \ldots, n_i}. Generally the 
covariate values from all treatments are stored together in one vector and 
\eqn{r = 1, \ldots, N} enumerates the values. For any integer between 1 and 
\eqn{N}, \code{mapindex()} tells which treatment the \eqn{r^{th}} value 
belongs to, and which observation in the identified treatment.

This is a rewrite of the original \code{mapindex()}function from the PSSIM
library using Rcpp to decrease runtime. If you load both this library and 
PSSIM, load this one second so the faster functions will overwrite the 
slower ones.
}
\examples{
r = 5; n = c(7, 8); mapindex(r, n)
r = 7; n = c(7, 8); mapindex(r, n)
r = 9; n = c(7, 8); mapindex(r, n)
 
}
\references{
Haiyan Wang, Siti Tolos, and Suojin Wang (2010). A Distribution Free
 Nonparametric Test to Detect Dependence Between a Response Variable and
 Covariate in Presence of Heteroscedastic Treatment Effects.
 The Canadian Journal of Statistics. 38(3), 408433. Doi:10.1002/cjs.10068
}
\seealso{
\code{\link[=NPtest_indept]{NPtest_indept}} uses this function to 
run the independence test.
}
