% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{NPtest_indept}
\alias{NPtest_indept}
\title{Test of independence in the presence of heteroscedastic treatments}
\usage{
NPtest_indept(dat, k = 7L)
}
\arguments{
\item{dat}{A dataframe with three columns named X, trt, and Y, where X is
the covariate, trt is the treatment level, and Y is the response variable.}

\item{k}{An odd integer to specify the number of nearest neighbors to be
used in augmentation. Generally recommend to use 3, 5, or 7.}
}
\value{
A list containing the following variables:
\itemize{
  \item \strong{Asys_var}: The asymptotic variance for the test statistics.
  \item \strong{Tstat}: The test statistic.
  \item \strong{pvalue}: The p-value of the test under \eqn{H_0}:
  There is independence between X and Y.
}
}
\description{
\code{NPtest_indept()} performs the test of independence between the 
response variable and a single covariate when there is potentially 
heteroscedastic treatment effects present (see Wang, Tolos and Wang (2010)).

This is a rewrite of the original \code{NPtest_indept()} function from the
PSSIM library using Rcpp to decrease runtime. If you load both this library 
and PSSIM, load this one second so the faster functions will overwrite the 
slower ones.
}
\examples{
n = 64;  X = runif(n); trt = gl(2, n/2)
e = rnorm(n, 0, 0.1)
Y = ifelse(trt == 1, 4*(X-0.5)^2+e, 2*X+e)
dat = data.frame(X, Y, trt)
NPtest_indept(dat, k = 7)
 
}
\references{
Haiyan Wang, Siti Tolos, and Suojin Wang (2010). A Distribution Free
 Nonparametric Test to Detect Dependence Between a Response Variable and
 Covariate in Presence of Heteroscedastic Treatment Effects.
 The Canadian Journal of Statistics. 38(3), 408433. Doi:10.1002/cjs.10068
}
\seealso{
\code{\link[=makepseudo]{makepseudo}}, 
\code{\link[=mapindex]{mapindex}}, and  
\code{\link[=trt_position]{trt_position}} are all called within this.
}
